requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118 
torch==2.0.1
torchvision==0.15.2 
torchaudio==2.0.2
diffusers==0.27.2
accelerate==0.28.0
tensorflow==2.12.0
tensorboard==2.12.0
opencv-python==4.9.0.80
soundfile==0.12.1
transformers==4.39.2

gdown
requests
imageio[ffmpeg]

omegaconf
ffmpeg-python
gradio
spaces
moviepy

flask==3.0.3
PyYAML==6.0.1
runpod==1.6.2

boto3
python-dotenv
tqdm